REST api producing complex synthetic data Based on 3d Perlin noise. So that I can retrieve a repeatable range of semi random complex data. Allowing me to experiment with different visualisation client tech stacks, leading to the creation of interesting browser based / server side applications that require control and visualisation of complex data models.
This data is represented as a JSON structure containing 2 main objects.
A 'header' block - used to detail the origin of the data, so that clients can request
additional data stepping forwards (look to adopt HATOS later)

A collection of data tupals. Each tupal is a conceptual x/y/z, or angle/value/time  dataset.

Where the header can be used to parses these tupals into a 3d matrix of vertex points.
Forming a contoured landscape. Since the data is generated by a mathematical alg. the
range of values can be tiled, or animated to simulate flying over a landscape or
flowing thru a liquid with variable density.

The data is conceptually a 3d cartesian system.
But can be adapted to form radial data patterns by simple interp. the axis as angle offsets
 and the z (height) being radius from centre.

The built in 'step' value of the REST request can be used to expand density of generated values
for when the radius become relatively large. So to introduce fine grained curved surface detailing.